create a folder to contain all sensor API
