+++ sensor_API/Source/Inc/TMAG5273.h
