+++ sensor_API/Source/Src/TMAG5273.cpp
